requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pycryptodome>=3.18.0
flask>=3.0.0
flask-cors>=4.0.0
